Checkin after refactoring
